--- a/docs/content/tools/_index.md
+++ b/docs/content/tools/_index.md
@@ -1,5 +1,4 @@
 ---
-lastmod: 2015-12-30
 date: 2015-09-12T10:40:31+02:00
 title: Tools
 weight: 120