--- a/hugolib/gitinfo.go
+++ b/hugolib/gitinfo.go
@@ -52,6 +52,7 @@ func (h *HugoSites) assembleGitInfo() {
 	s := h.Sites[0]
 
 	for _, p := range s.AllPages {
+		// TODO(bep) np consider other nodes
 		// Git normalizes file paths on this form:
 		filename := path.Join(contentRoot, contentDir, filepath.ToSlash(p.Path()))
 		g, ok := gitMap[filename]