… docs/content/content/multilingual.md
… docs/content/content/sections.md
… docs/content/content/summaries.md
… docs/content/extras/menus.md
… docs/content/extras/pagination.md
… docs/content/extras/scratch.md
… docs/content/meta/release-notes.md
… docs/content/overview/source-directory.md
… docs/content/taxonomies/usage.md
… docs/content/templates/404.md
… docs/content/templates/debugging.md
… docs/content/templates/functions.md
… docs/content/templates/go-templates.md
… docs/content/templates/homepage.md
… docs/content/templates/list.md
… docs/content/templates/partials.md
… docs/content/templates/rss.md
… docs/content/templates/sitemap.md
… docs/content/templates/terms.md
… docs/content/templates/variables.md
… docs/layouts/_default/list.html
… hugolib/embedded_shortcodes_test.go
… hugolib/hugo_sites_build.go
… hugolib/hugo_sites_build_test.go
… hugolib/hugo_sites_test.go
… hugolib/menu.go
… hugolib/menu_test.go
… hugolib/node.go
… hugolib/node_as_page_test.go
… hugolib/node_test.go