- [x] Basic test and types
- [x] Handle home page
- [x] Handle sections
- [x] Handle taxonomy lists
- [x] Handle taxonomy terms
- [x] Create pages for missing content files for sections etc.
- [x] Get Rebuild up to speed
- [x] Taxonomies in home page etc?
- [x] Content and page metadata  in paginator pages? Fix in #2682
- [x] Expose a `GetPage(NodeType, path...)` on `SiteInfo` or similar ...
- [x] Handle aliases
- [x] Handle menus
- [x] Menu definitions in home page etc?
- [x] Alias definitions in home page etc?
- [x] Handle RSS
- [x] Handle sitemaps
- [x] Handle sections with only `_index.md`
- [x] Include all node types in sitemap template, fixes #1303 
- [x] Check Date, Lastmod on nodes
- [x] Drafts etc.
- [x] Handle 404
- [x] Handle robots.txt
- [x] Handle translations
- [x]  Fix the All* slices, add others? `AllNodes`?
- [x] Check `where` vs `NodeType` 
- [x] Add NodeContent path setting default to /Content Fix in #2680
- [x] Node Content from themes? Fix in #2681
- [x] Do the `TODO(bep) np`
- [x] Remove old and dead code
- [x] Remove Nodes, AllNodes (=> Pages), NodeType => PageType, PageType = string, so we can query it: home, section, taxonomy, taxonomyTerm. Data.Pages => .Pages and a mild deprecation (no ERROR)
- [x] Rename `PageType`=> `Kind`
- [x] Update RSSlink => RSSLink in themes and test the theme site
- [x] Compare /docs site before/after
- [x] Check sitemap sorting / stability
- [x] Don't write taxonomy terms page if no layout for it / i.e. how it is today
- [x]  Docs site: /tools/index.html is rendered empty
- [x] Check multilingual `_index.md`
- [x] Do Docs
- [x] Add some lines to the release notes while it is fresh
- [x] Check theme templates for sections etc.
++

Fixes #2297