@bep 2016-11-09
The tests are all green -- and this actually works now. Still many loose ends to tie, but it looks promising.

**And faster:**

```
benchmark           old ns/op       new ns/op       delta
BenchmarkHugo-4     12606861551     11300837482     -10.36%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     48522596       48506418       -0.03%

benchmark           old bytes      new bytes      delta
BenchmarkHugo-4     9800844224     9791154400     -0.10%
```