@bep 2016-11-14
Removing the `Node` had further positive impact on speed, I believe the value embedding created some extra copies.

This compares Hugo `master` with this branch:
```
benchmark           old ns/op       new ns/op       delta
BenchmarkHugo-4     12763741678     10800430319     -15.38%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     48522867       42625712       -12.15%

benchmark           old bytes      new bytes      delta
BenchmarkHugo-4     9796458336     9492564016     -3.10%
```