@bep 2016-11-17
I will add some docs, but it would be nice if someone would review this. 

This is a big change, so when looking this it would nice if you

* Look at the big lines and obvious mistakes / bugs
* This is an incremental step, so there are things that I would have liked to done nicer, but I had to stop at some point ...
* Also, if you want to express new feature request, do that in separate issues

/cc @digitalcraftsman @spf13 @moorereason @anthonyfok