@moorereason 2016-11-17
I did a quick comparison of Hugo docs site output vs master.

### RSS XML

The RSS feed Title and Description are duped now that the homepage is a Page.  Template should check `.IsHome`?

```diff
diff -urN public-master/index.xml public-pages/index.xml
--- public-master/index.xml 2016-11-17 08:42:11.660808150 -0600
+++ public-pages/index.xml  2016-11-17 08:44:42.025685195 -0600
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8" standalone="yes" ?>
 <rss version="2.0" xmlns:atom="http://www.w3.org/2005/Atom">
   <channel>
-    <title>Hugo: A Fast and Flexible Website Generator</title>
+    <title>Hugo: A Fast and Flexible Website Generator on Hugo: A Fast and Flexible Website Generator</title>
     <link>http://gohugo.io/index.xml</link>
-    <description>Recent content on Hugo: A Fast and Flexible Website Generator</description>
+    <description>Recent content in Hugo: A Fast and Flexible Website Generator on Hugo: A Fast and Flexible Website Generator</description>
     <generator>Hugo -- gohugo.io</generator>
     <lastBuildDate>Thu, 08 Jan 2015 16:11:23 -0700</lastBuildDate>
     <atom:link href="http://gohugo.io/index.xml" rel="self" type="application/rss+xml" />
```

### Minified index pages

It looks like most of the index.html pages are now minified (rendered onto a single line).  For an example, see `meta/release-notes/index.html`.  Was this done on purpose?

### lastBuildDate in Taxonomy RSS

Looks like we're now getting `lastBuildDate` on all Taxonomy RSS where it used to be left off.  For an example, see `tags/blog/index.xml`.  It looks like this is working fine, but I wanted to point out the addition.