@moorereason 2016-11-17
Tested building `github.com/bep/hugo-benchmark/sites/ado-hugo` and several pages are not created:

- about/index.html
- page/2/index.html
- page/3/index.html
- page/4/index.html
- page/5/index.html
- page/6/index.html
- page/7/index.html
- page/8/index.html