@bep 2016-11-17
@moorereason thanks for the heads-up on  the ado-hugo, that looks funky; re the diffs, I'll make a note about adjusting the RSS template, but I should point out that I would not expect exact the same as before, that would not be fruitful. Earlier the node creation was all over the place.