@bep 2016-11-17
@moorereason I forgot to append the theme templates when rendering home page etc... which was a pretty big breaking change for some ... hehe. Fixed now, and luckily the benchmark is still on the positive side:

```
benchmark           old ns/op       new ns/op       delta
BenchmarkHugo-4     12081022610     10461159669     -13.41%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     47515450       41721534       -12.19%

benchmark           old bytes      new bytes      delta
BenchmarkHugo-4     9237288600     8870780552     -3.97%
```