@moorereason 2016-11-17
Regarding the minified pages I mentioned before, I found the issue but not the root cause.  The "minified" result is actually the internal Alias template from `target/htmlredirect.go`.  I'm not sure why it's happening, but I see it with Hugo docs `meta/release-notes/index.html` which redirects to `http://gohugo.io/meta/release-notes/` and with ado-hugo `gpm/index.html` which redirects to `https://www.arresteddevops.com/gpm`.  In both cases, the expected page content is never actually written to disk (no new file with all the content is created anywhere).