@bep 2016-11-18
> but I see it with Hugo docs meta/release-notes/index.html which redirects to http://gohugo.io/meta/release-notes/ and

Are you sure you run with this PR branch (and an up-to-date one)? Because I cannot reproduce the release notes redirect nor the RSS home page title/description issue you mentioned.