@moorereason 2016-11-18
I double-checked everything, and yes, I'm using the latest PR.
```
$ hugo-pages version
Hugo Static Site Generator v0.18-DEV-93D5EF2 BuildDate: 2016-11-18T08:19:08-06:00
```

```
$ git log -1
commit 93d5ef29cf51fd181ad45ccee357180f96406749
Author: Bjørn Erik Pedersen <bjorn.erik.pedersen@gmail.com>
Date:   Thu Nov 17 17:08:31 2016 +0100

     node to page: Need to append theme templates

     Updates #2297
```

I've also run `govendor sync github.com/spf13/hugo` to make sure I have the correct vendor packages.  I'm generating the docs site within the PR branch.