@moorereason 2016-11-18
We should make the sitemap.xml sorting stable.