@bep 2016-11-18
OK - I will take it for a spin on Windows and see if that breaks.