@bep 2016-11-18
>FWIW, I'm testing with go1.7.3 linux/amd64.

On Windows?