@moorereason 2016-11-18
No, I'm not in Windows.  I rarely ever use Windows with hugo unless someone reports a problem requiring its use.  I'm in Arch Linux.