@bep 2016-11-19
@moorereason I can reproduce it on OSX, too. Thanks for the heads up. I get the "mistakes" in different files than you, which looks like a ordering issue. Looks like the real pages gets overwritten by an alias refresh. But that means that I can debug and fix it.