@bep 2016-11-19
I have fixed the basic issue in 35e8cc870ab49b66d19724b7ab771be26f5e616e -- I will have a look at some other minor diffs in master/docs vs this/docs later.