@bep 2016-11-20
@moorereason now the /docs before/after is down to 3 diffs:

```
Files /Users/bep/hugo_master/index.html and docs/public/index.html differ
Files /Users/bep/hugo_master/sitemap.xml and docs/public/sitemap.xml differ
Files /Users/bep/hugo_master/tools/index.html and docs/public/tools/index.html differ
```

1. Is because of the `shuffle` usage and is OK
2. We now output nodes + pages. There is a potential sorting issue with nodes, but not about stability in my head. I have noted at TODO about it, not sure I'm going to fix it now.
3. Docs has both a page and a section that ends up in /tools. So this is a ordering issue that I'm not sure how/if I'm going to fix. It is visible with the new code because we do more in parallel.