@bep 2016-11-21
I consider this to be ready to merge now, and will do so if I hear no protests in the next two days. This is tested on the border of paranoia, but I'm pretty sure that with changes on this scale, there will be issues in the wild that only the wild can tell me about.

I have benched this against v0.17, and the accumulated changes is suspiciously good, so I will rerun this test at some point:
 
```bash
benchmark           old ns/op       new ns/op       delta
BenchmarkHugo-4     26227496901     10160470393     -61.26%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     73049239       41718032       -42.89%

benchmark           old bytes       new bytes      delta
BenchmarkHugo-4     11190001512     8875229560     -20.69%
```