@bep 2016-11-21
A benchmark for this branch vs master:

```
benchmark           old ns/op       new ns/op       delta
BenchmarkHugo-4     12754569495     10079717966     -20.97%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     47509425       41718139       -12.19%

benchmark           old bytes      new bytes      delta
BenchmarkHugo-4     9257774288     8880549424     -4.07%
```